change way of selecting a city
In this commit i have provided another manner of picking the city.
Instead of fetching all the possible cities and fill the select
box with them I decided to use typeahead input and getting
Only those of cities that includes a substring of at least 2 letters
typed by the user. It significantly improve working of my application.
I have get rid of the lag caused by very huge amount of data
that have been brought by this previous request to the database.